+++ docs/src/conf.py
@@ -29,7 +29,7 @@ author = "ATLAS Team"
 
 # The full version, including alpha/beta/rc tags
 
-release = "31.2.5-test-upgradability-b"
+release = "31.4.0"
 
 # -- General configuration ---------------------------------------------------